flask-app/requirements.txt
kagglehub
flask==2.3.3
werkzeug==2.3.7
tensorflow==2.13.1
numpy==1.24.3
kubernetes==26.1.0
python-logging-loki==0.3.1
typing-extensions>=3.6.6,<4.6.0
requests>=2.28.0
